requirements.txt
jinja2==3.0.3
folium==0.14.0
branca==0.6.0
python-dotenv==1.0.0
black==23.7.0
gradio==3.42.0
leafmap==0.23.4
pandas==2.0.3
numpy==1.24.4
geopandas==0.13.2
ipython==7.28.0
watchdog==2.1.6
fastapi==0.68.1
pydantic==1.8.2
loguru==0.7.2
tomark==0.1.4
polyline==2.0.1
colorcet==3.0.1
webcolors==1.13.0